# Conflicts:
#	app/assets/javascripts/discourse-common/package.json
#	app/assets/javascripts/discourse-plugins/package.json
#	app/assets/javascripts/discourse/package.json
#	app/assets/javascripts/package.json
#	app/assets/javascripts/yarn.lock
#	app/views/qunit/theme.html.erb
#	lib/discourse_js_processor.rb
#	lib/ember_cli.rb
#	lib/freedom_patches/barber_ember_precompiler.rb
#	lib/freedom_patches/raw_handlebars.rb
#	lib/tasks/docker.rake
#	lib/tasks/javascript.rake
#	lib/tasks/svg_icons.rake
#	package.json
#	spec/lib/discourse_js_processor_spec.rb
#	yarn.lock

+++ app/models/draft.rb
@@ -1,21 +1,23 @@
 # frozen_string_literal: true
 
 class Draft < ActiveRecord::Base
-  NEW_TOPIC ||= 'new_topic'
-  NEW_PRIVATE_MESSAGE ||= 'new_private_message'
-  EXISTING_TOPIC ||= 'topic_'
+  NEW_TOPIC ||= "new_topic"
+  NEW_PRIVATE_MESSAGE ||= "new_private_message"
+  EXISTING_TOPIC ||= "topic_"
 
   belongs_to :user
 
-  after_commit :update_draft_count, on: [:create, :destroy]
+  after_commit :update_draft_count, on: %i[create destroy]
 
-  class OutOfSequence < StandardError; end
+  class OutOfSequence < StandardError
+  end
 
   def self.set(user, key, sequence, data, owner = nil, force_save: false)
     return 0 if !User.human_user_id?(user.id)
     force_save = force_save.to_s == "true"
 
-    if SiteSetting.backup_drafts_to_pm_length > 0 && SiteSetting.backup_drafts_to_pm_length < data.length
+    if SiteSetting.backup_drafts_to_pm_length > 0 &&
+         SiteSetting.backup_drafts_to_pm_length < data.length
       backup_draft(user, key, sequence, data)
     end
 
@@ -44,21 +46,16 @@ class Draft < ActiveRecord::Base
     current_sequence ||= 0
 
     if draft_id
-      if !force_save && (current_sequence != sequence)
-        raise Draft::OutOfSequence
-      end
+      raise Draft::OutOfSequence if !force_save && (current_sequence != sequence)
 
       sequence = current_sequence if force_save
       sequence += 1
 
       # we need to keep upping our sequence on every save
       # if we do not do that there are bad race conditions
-      DraftSequence.upsert({
-          sequence: sequence,
-          draft_key: key,
-          user_id: user.id,
-        },
-        unique_by: [:user_id, :draft_key]
+      DraftSequence.upsert(
+        { sequence: sequence, draft_key: key, user_id: user.id },
+        unique_by: %i[user_id draft_key],
       )
 
       DB.exec(<<~SQL, id: draft_id, sequence: sequence, data: data, owner: owner || current_owner)
@@ -70,17 +67,10 @@ class Draft < ActiveRecord::Base
              , updated_at = CURRENT_TIMESTAMP
          WHERE id = :id
       SQL
-
     elsif sequence != current_sequence
       raise Draft::OutOfSequence
     else
-      opts = {
-        user_id: user.id,
-        draft_key: key,
-        data: data,
-        sequence: sequence,
-        owner: owner
-      }
+      opts = { user_id: user.id, draft_key: key, data: data, sequence: sequence, owner: owner }
 
       draft_id = DB.query_single(<<~SQL, opts).first
         INSERT INTO drafts (user_id, draft_key, data, sequence, owner, created_at, updated_at)
@@ -101,8 +91,8 @@ class Draft < ActiveRecord::Base
 
     UploadReference.ensure_exist!(
       upload_ids: Upload.extract_upload_ids(data),
-      target_type: 'Draft',
-      target_id: draft_id
+      target_type: "Draft",
+      target_id: draft_id,
     )
 
     sequence
@@ -111,11 +101,7 @@ class Draft < ActiveRecord::Base
   def self.get(user, key, sequence)
     return if !user || !user.id || !User.human_user_id?(user.id)
 
-    opts = {
-      user_id: user.id,
-      draft_key: key,
-      sequence: sequence
-    }
+    opts = { user_id: user.id, draft_key: key, sequence: sequence }
 
     current_sequence, data, draft_sequence = DB.query_single(<<~SQL, opts)
       WITH draft AS (
@@ -136,9 +122,7 @@ class Draft < ActiveRecord::Base
 
     current_sequence ||= 0
 
-    if sequence != current_sequence
-      raise Draft::OutOfSequence
-    end
+    raise Draft::OutOfSequence if sequence != current_sequence
 
     data if current_sequence == draft_sequence
   end
@@ -155,9 +139,7 @@ class Draft < ActiveRecord::Base
     current_sequence = DraftSequence.current(user, key)
 
     # bad caller is a reason to complain
-    if sequence != current_sequence
-      raise Draft::OutOfSequence
-    end
+    raise Draft::OutOfSequence if sequence != current_sequence
 
     # corrupt data is not a reason not to leave data
     Draft.where(user_id: user.id, draft_key: key).destroy_all
@@ -176,9 +158,7 @@ class Draft < ActiveRecord::Base
   end
 
   def topic_id
-    if draft_key.starts_with?(EXISTING_TOPIC)
-      draft_key.gsub(EXISTING_TOPIC, "").to_i
-    end
+    draft_key.gsub(EXISTING_TOPIC, "").to_i if draft_key.starts_with?(EXISTING_TOPIC)
   end
 
   def topic_preloaded?
@@ -186,7 +166,9 @@ class Draft < ActiveRecord::Base
   end
 
   def topic
-    topic_preloaded? ? @topic : @topic = Draft.allowed_draft_topics_for_user(user).find_by(id: topic_id)
+    topic_preloaded? ?
+      @topic :
+      @topic = Draft.allowed_draft_topics_for_user(user).find_by(id: topic_id)
   end
 
   def preload_topic(topic)
@@ -240,10 +222,7 @@ class Draft < ActiveRecord::Base
     offset = (opts[:offset] || 0).to_i
     limit = (opts[:limit] || 30).to_i
 
-    stream = Draft.where(user_id: user_id)
-      .order(updated_at: :desc)
-      .offset(offset)
-      .limit(limit)
+    stream = Draft.where(user_id: user_id).order(updated_at: :desc).offset(offset).limit(limit)
 
     # Preload posts and topics to avoid N+1 queries
     Draft.preload_data(stream, opts[:user])
@@ -273,16 +252,12 @@ class Draft < ActiveRecord::Base
     reply = JSON.parse(data)["reply"] || ""
     return if reply.length < SiteSetting.backup_drafts_to_pm_length
 
-    post_id = BackupDraftPost.where(user_id: user.id, key: key).pluck_first(:post_id)
+    post_id = BackupDraftPost.where(user_id: user.id, key: key).pick(:post_id)
     post = Post.where(id: post_id).first if post_id
 
-    if post_id && !post
-      BackupDraftPost.where(user_id: user.id, key: key).delete_all
-    end
+    BackupDraftPost.where(user_id: user.id, key: key).delete_all if post_id && !post
 
-    indented_reply = reply.split("\n").map! do |l|
-      "    #{l}"
-    end
+    indented_reply = reply.split("\n").map! { |l| "    #{l}" }
     draft_body = <<~MD
       #{indented_reply.join("\n")}
 
@@ -297,13 +272,14 @@ class Draft < ActiveRecord::Base
     if !post
       topic = ensure_draft_topic!(user)
       Post.transaction do
-        post = PostCreator.new(
-          user,
-          raw: draft_body,
-          skip_jobs: true,
-          skip_validations: true,
-          topic_id: topic.id,
-        ).create
+        post =
+          PostCreator.new(
+            user,
+            raw: draft_body,
+            skip_jobs: true,
+            skip_validations: true,
+            topic_id: topic.id,
+          ).create
         BackupDraftPost.create!(user_id: user.id, key: key, post_id: post.id)
       end
     elsif post.last_version_at > 5.minutes.ago
@@ -311,48 +287,47 @@ class Draft < ActiveRecord::Base
       post.update_columns(
         raw: draft_body,
         cooked: PrettyText.cook(draft_body),
-        updated_at: Time.zone.now
+        updated_at: Time.zone.now,
       )
     else
       revisor = PostRevisor.new(post, post.topic)
-      revisor.revise!(user, { raw: draft_body },
+      revisor.revise!(
+        user,
+        { raw: draft_body },
         bypass_bump: true,
         skip_validations: true,
         skip_staff_log: true,
-        bypass_rate_limiter: true
+        bypass_rate_limiter: true,
       )
     end
-
   rescue => e
     Discourse.warn_exception(e, message: "Failed to backup draft")
   end
 
   def self.ensure_draft_topic!(user)
-    topic_id = BackupDraftTopic.where(user_id: user.id).pluck_first(:topic_id)
+    topic_id = BackupDraftTopic.where(user_id: user.id).pick(:topic_id)
     topic = Topic.find_by(id: topic_id) if topic_id
 
-    if topic_id && !topic
-      BackupDraftTopic.where(user_id: user.id).delete_all
-    end
+    BackupDraftTopic.where(user_id: user.id).delete_all if topic_id && !topic
 
     if !topic
       Topic.transaction do
-        creator = PostCreator.new(
-          user,
-          title: I18n.t("draft_backup.pm_title"),
-          archetype: Archetype.private_message,
-          raw: I18n.t("draft_backup.pm_body"),
-          skip_jobs: true,
-          skip_validations: true,
-          target_usernames: user.username
-        )
+        creator =
+          PostCreator.new(
+            user,
+            title: I18n.t("draft_backup.pm_title"),
+            archetype: Archetype.private_message,
+            raw: I18n.t("draft_backup.pm_body"),
+            skip_jobs: true,
+            skip_validations: true,
+            target_usernames: user.username,
+          )
         topic = creator.create.topic
         BackupDraftTopic.create!(topic_id: topic.id, user_id: user.id)
       end
     end
 
     topic
-
   end
 
   def update_draft_count
